Riz
(auto-msg whatthecommit.com)

On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	modified:   src/parser/mod.rs
	modified:   src/parser/types.rs

+++ src/parser/types.rs
@@ -116,14 +116,14 @@ pub struct BinaryExpression {
 pub struct FloatLiteral {
     pub coin: Coin<String>,
     pub ttype: Lexeme,
-    pub val: f32,
+    pub val: f64,
 }
 
 #[derive(Debug, PartialEq,Clone)]
 pub struct IntLiteral {
     pub coin: Coin<String>,
     pub ttype: Lexeme,
-    pub val: i64,
+    pub val: u64,
 }
 
 #[derive(Debug, PartialEq,Clone)]
